adds osrm distance implementation

+++ src/main/java/com/github/foodie/services/DistanceService.java
@@ -2,6 +2,8 @@ package com.github.foodie.services;
 
 import org.locationtech.jts.geom.Point;
 
+import java.math.BigDecimal;
+
 public interface DistanceService {
-    double calculateDistance(Point src, Point dest);
+    BigDecimal calculateDistance(Point src, Point dest);
 }
